requirements.txt
altgraph==0.17.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
idna==3.11
macholib==1.16.4
packaging==26.0
pillow==12.1.1
PyAudio==0.2.14
pycparser==3.0
pyinstaller @ git+https://github.com/pyinstaller/pyinstaller.git@42c16f919ce9be424fd33ce926a2bbce05c42430
pyinstaller-hooks-contrib==2026.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
requests==2.32.5
srt==3.5.3
tqdm==4.67.3
urllib3==2.6.3
vosk==0.3.44
websockets==16.0
